% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningRcppShark.R
\docType{package}
\name{LearningRcppShark-package}
\alias{LearningRcppShark}
\alias{LearningRcppShark-package}
\title{LearningRcppShark: R Bindings To Shark ML Algorithms}
\description{
This is my attempt to learn Shark C++ machine learning library
via the RcppShark package.
}

