% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{predict.SharkKM}
\alias{predict.SharkKM}
\title{Predict using a k-means}
\usage{
\method{predict}{SharkKM}(obj, newdata, ...)
}
\arguments{
\item{newdata}{A \code{matrix} to classify}

\item{...}{Further arguments passed to or from other methods}

\item{x}{A \code{SharkKM} model object}
}
\examples{
y <- shark_kmeans(iris[, 1:4], 3)
y_hat <- predict(y, iris[, 1:4])
}

