% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{shark_kmeans}
\alias{shark_kmeans}
\title{Train a k-means model}
\usage{
shark_kmeans(x, k)
}
\arguments{
\item{x}{A \code{matrix} of observations to cluster}

\item{k}{Number of clusters}
}
\examples{
y <- shark_kmeans(iris[, 1:4], 3)
}

